+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Add feedback to user when updating wallet watch-only settings
 - Enable jade channel selector via command line option
 - Increased the size of sign up dialog titles
 - Moved empty transactions list label to top left of the container